fix Makefile bug dont need bal
