+++ am/arch/mips32-npc/img/boot/Makefile
@@ -1,2 +1,2 @@
 start.o: start.S
-	mips-linux-gnu-as -EL -march=mips32 start.S -o start.o
+	mips-linux-gnu-gcc -EL -march=mips32 -fno-pic -MD -mno-abicalls -fno-delayed-branch -c start.S -o start.o
