+++ am/arch/mips32-npc/src/trm.c
@@ -25,12 +25,22 @@ void _halt(int code) {
   while(1);
 }
 
+u32 GetCount(){
+  u32 tick = 0;
+  asm volatile("mfc0 %0, $9\n\t":"=r"(tick));
+  return tick;
+}
+
+void SetCompare(u32 compare){
+  asm volatile("mtc0 %0, $11\n\t"::"r"(compare));
+}
+
 ulong _uptime() {
   return npc_time ++;
 }
 
 ulong _cycles(){
-  return npc_cycles ++;
+  return 0;
 }
 
 void memory_init(){
